Add types for SystemJS@6
