+++ types/systemjs/systemjs-tests.ts
@@ -1,68 +1,78 @@
-import SystemJS = require('systemjs');
-
-SystemJS.config({
-    baseURL: '/app'
+System.import('./hi.js').then((hi: SystemJS.Module) => {
+    hi.someProperty();
 });
 
-SystemJS.import('main.js');
-
-SystemJS.config({
-    // 'plugin-traceur' or 'plugin-typescript' or 'babel' or 'traceur' or 'typescript' or false.
-    transpiler: 'plugin-babel',
-    // or traceurOptions or typescriptOptions
-    babelOptions: {
-    }
+System.import<Hi>('./hi.js').then(hi => {
+    hi.someExport();
 });
 
-SystemJS.config({
-    map: {
-        traceur: 'path/to/traceur.js'
-    }
-});
+System.import('./hi.js', 'https://example.com/base/');
 
-SystemJS.config({
-    meta: {
-        '*': {
-            authorization: true
+System.register(['foo', 'bar'], (_export, _context) => {
+    let foo;
+    let bar;
+
+    return {
+        setters: [
+            module => {
+                foo = module;
+            },
+            module => {
+                bar = module;
+            }
+        ],
+        execute() {
+            _export('a', 'thing');
+            _export('b', 123);
+            _export('c', () => 'hi');
+            _export({some: 'thing'});
+
+            _context.import('./other-thing.js');
+
+            _context.meta.url;
         }
-    }
+    };
 });
 
-SystemJS.config({
-    meta: {
-        '*': {
-            authorization: 'Basic YWxhZGRpbjpvcGVuc2VzYW1l'
-        }
-    }
+// named register
+System.register('name', [], () => ({}));
+
+const update = System.delete('https://example.com/a.js');
+if (update) {
+    update();
+} else {
+    const expected: false = update;
+}
+
+const a = System.get('https://example.com/a.js');
+if (a) {
+    a.doThing();
+} else {
+    const expected: null = a;
+}
+
+const b = System.get<ModuleB>('https://example.com/b.js');
+if (b) {
+    b.theBThing();
+} else {
+    const expected: null = b;
+}
+
+const hasC: boolean = System.has('https://example.com/c.js');
+
+System.set('https://example.com/d.js', {
+    hi: 'there'
 });
 
-SystemJS.config({
-    map: {
-        'local/package': {
-            x: 'vendor/x.js'
-        },
-        'another/package': {
-            x: 'vendor/y.js'
-        }
-    }
-});
+for (const entry of System.entries()) {
+    const moduleId: string = entry[0];
+    const module: SystemJS.Module = entry[1];
+}
 
-SystemJS.transpiler = 'traceur';
+interface ModuleB {
+    theBThing(): void;
+}
 
-const mockModule = {
-    default: () => {
-        return 42;
-    }
-};
-
-SystemJS.set('./app.js', SystemJS.newModule(mockModule));
-
-SystemJS.import('./app.js').then((m: typeof mockModule) => {
-    m.default();
-});
-
-SystemJS.import('lodash').then((_: (...args: any[]) => any) => {
-    _(1, '2', {}, []);
-});
-
-const clonedSystemJSJS = new SystemJS.constructor();
+interface Hi {
+    someExport(): void;
+}
