+++ types/systemjs/tslint.json
@@ -1,3 +1,8 @@
 {
-  "extends": "dtslint/dt.json"
+  "extends": "dtslint/dt.json",
+  "rules": {
+    "no-unnecessary-generics": {
+      "severity": "off"
+    }
+  }
 }
